Remove CheckValid() anywhere an object already exists, as it is assumed valid already
we really only need to check for subgroup on point's where keys are first constructed.

I noticed that the check isn't cheap, the
 if(pubKey.CheckValid()) {
        g1_copy(q, pubKey.q);
    }

caused bench to slow down 10x for batch verifications. If we can assume PublicKey being passed in would have already been verified we don't need this check.