Use system Catch2 library instead of bundled one if found.
Bundled Catch2 library (2.13.4) is outdated and doesn't compile with
glibc 2.34 because MINSIGSTKSZ is not a constant anymore.

Fortunately, this problem is fixed in newer version of Catch2,
and modern Linux distros (like Fedora) have this library in its repos.

This change makes cmake check for Catch2 provided by the system and
use it instead of bundled one if found.

Also, it adds include directories needed by "catch.hpp" only to
runtest target because it's used only there.

Signed-off-by: Oleg Girko <ol@infoserver.lv>